+++ src/main/java/de/maxhenkel/easyvillagers/entity/EasyVillagerEntity.java
@@ -1,0 +1,18 @@
+package de.maxhenkel.easyvillagers.entity;
+
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.merchant.villager.VillagerEntity;
+import net.minecraft.entity.villager.VillagerType;
+import net.minecraft.world.World;
+
+public class EasyVillagerEntity extends VillagerEntity {
+
+    public EasyVillagerEntity(EntityType<? extends VillagerEntity> type, World worldIn) {
+        super(type, worldIn);
+    }
+
+    public EasyVillagerEntity(EntityType<? extends VillagerEntity> type, World worldIn, VillagerType villagerType) {
+        super(type, worldIn, villagerType);
+    }
+
+}